requirements.txt
exif
pyexiftool
moviepy